changed decimal to float in recipe_ingredient migration

+++ database/migrations/2019_01_15_094057_create_ingredient_recipe_table.php
@@ -17,7 +17,7 @@ class CreateIngredientRecipeTable extends Migration
             $table->increments('id');
             $table->integer('ingredient_id')->unsigned();
             $table->integer('recipe_id')->unsigned();
-            $table->decimal('amount',8,3);
+            $table->float('amount',8,2);
             $table->char('unit',10);
 
             $table->foreign('ingredient_id')->references('id')->on('ingredients');;
